// frontend/src/services/OpenApi/models/ForestClientDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CodeDescriptionDto } from './CodeDescriptionDto';
export type ForestClientDto = {
    clientNumber: string;
    clientName: string;
    legalFirstName: string;
    legalMiddleName: string;
    clientStatusCode: CodeDescriptionDto;
    clientTypeCode: CodeDescriptionDto;
    acronym: string;
    name?: string;
};